Move generated data to builder folder
