+++ builder/amazon/common/interpolate_build_info.go
@@ -3,7 +3,7 @@ package common
 import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	builderscommon "github.com/hashicorp/packer/builder/common"
+	"github.com/hashicorp/packer/builder"
 	"github.com/hashicorp/packer/helper/multistep"
 )
 
@@ -16,7 +16,7 @@ type BuildInfoTemplate struct {
 	SourceAMITags      map[string]string
 }
 
-func extractBuildInfo(region string, state multistep.StateBag, generatedData *builderscommon.GeneratedData) *BuildInfoTemplate {
+func extractBuildInfo(region string, state multistep.StateBag, generatedData *builder.GeneratedData) *BuildInfoTemplate {
 	rawSourceAMI, hasSourceAMI := state.GetOk("source_image")
 	if !hasSourceAMI {
 		return &BuildInfoTemplate{
