+++ builder/amazon/common/step_modify_ami_attributes.go
@@ -3,11 +3,10 @@ package common
 import (
 	"context"
 	"fmt"
-	builderscommon "github.com/hashicorp/packer/builder/common"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ec2"
+	"github.com/hashicorp/packer/builder"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/template/interpolate"
@@ -22,7 +21,7 @@ type StepModifyAMIAttributes struct {
 	Description    string
 	Ctx            interpolate.Context
 
-	GeneratedData *builderscommon.GeneratedData
+	GeneratedData *builder.GeneratedData
 }
 
 func (s *StepModifyAMIAttributes) Run(ctx context.Context, state multistep.StateBag) multistep.StepAction {
