+++ builder/amazon/ebssurrogate/builder.go
@@ -9,11 +9,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	builderscommon "github.com/hashicorp/packer/builder/common"
-
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/iam"
 	"github.com/hashicorp/hcl/v2/hcldec"
+	"github.com/hashicorp/packer/builder"
 	awscommon "github.com/hashicorp/packer/builder/amazon/common"
 	"github.com/hashicorp/packer/common"
 	"github.com/hashicorp/packer/helper/communicator"
@@ -178,7 +177,7 @@ func (b *Builder) Run(ctx context.Context, ui packer.Ui, hook packer.Hook) (pack
 	state.Put("awsSession", session)
 	state.Put("hook", hook)
 	state.Put("ui", ui)
-	generatedData := &builderscommon.GeneratedData{State: state}
+	generatedData := &builder.GeneratedData{State: state}
 
 	var instanceStep multistep.Step
