+++ builder/amazon/instance/builder.go
@@ -9,7 +9,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	builderscommon "github.com/hashicorp/packer/builder/common"
+	"github.com/hashicorp/packer/builder"
 	"os"
 	"strings"
 
@@ -249,7 +249,7 @@ func (b *Builder) Run(ctx context.Context, ui packer.Ui, hook packer.Hook) (pack
 	state.Put("awsSession", session)
 	state.Put("hook", hook)
 	state.Put("ui", ui)
-	generatedData := &builderscommon.GeneratedData{State: state}
+	generatedData := &builder.GeneratedData{State: state}
 
 	var instanceStep multistep.Step
